move PureModelContextDataGenerator to compiler
