+++ legend-engine-xts-analytics/legend-engine-xts-analytics-function/legend-engine-xt-analytics-function-api/pom.xml
@@ -38,10 +38,6 @@
         <!-- ENGINE -->
         <dependency>
             <groupId>org.finos.legend.engine</groupId>
-            <artifactId>legend-engine-external-shared</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.finos.legend.engine</groupId>
             <artifactId>legend-engine-language-pure-modelManager</artifactId>
         </dependency>
         <dependency>
