+++ legend-engine-xts-data-space/legend-engine-xt-data-space-generation/pom.xml
@@ -101,10 +101,6 @@
         </dependency>
         <dependency>
             <groupId>org.finos.legend.engine</groupId>
-            <artifactId>legend-engine-external-shared</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.finos.legend.engine</groupId>
             <artifactId>legend-engine-xt-analytics-mapping-protocol</artifactId>
         </dependency>
         <dependency>
