+++ legend-engine-xts-diagram/legend-engine-xt-diagram-api/pom.xml
@@ -58,10 +58,6 @@
         </dependency>
         <dependency>
             <groupId>org.finos.legend.engine</groupId>
-            <artifactId>legend-engine-external-shared</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.finos.legend.engine</groupId>
             <artifactId>legend-engine-xt-diagram-compiler</artifactId>
         </dependency>
         <!-- ENGINE -->
